Fix Windows dtype mismatch: use long long (int64) for COO matrix indices
On Windows, 'long' is 32-bit while on Linux/macOS it's 64-bit.
Use 'long long' (int64) explicitly for scipy COO matrix indices to ensure
cross-platform compatibility.

Fixes: ValueError: Buffer dtype mismatch, expected 'long' but got 'long long'

+++ implicit/_nearest_neighbours.pyx
@@ -131,8 +131,8 @@ def all_pairs_knn(users, unsigned int K=100, int num_threads=0, show_progress=Tr
 
     # holds triples of output
     cdef double[:] values = np.zeros(item_count * K)
-    cdef long[:] rows = np.zeros(item_count * K, dtype=int)
-    cdef long[:] cols = np.zeros(item_count * K, dtype=int)
+    cdef long long[:] rows = np.zeros(item_count * K, dtype=np.int64)
+    cdef long long[:] cols = np.zeros(item_count * K, dtype=np.int64)
 
     progress = tqdm(total=item_count, disable=not show_progress)
     with nogil, parallel(num_threads=num_threads):
